added problem, where spring context or just database is initialized for every springprofile

+++ test/src/main/java/org/rvanooyen/spring/test/repositories/ExampleRepository.java
@@ -3,5 +3,5 @@ package org.rvanooyen.spring.test.repositories;
 import org.rvanooyen.spring.test.model.ExampleModel;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ExampleRepository extends CrudRepository<ExampleModel, Long> {
+public interface ExampleRepository extends CrudRepository<ExampleModel, Integer> {
 }
